Image processed

+++ src/daos/restaurant.ts
@@ -2,6 +2,8 @@ import { Restaurant } from '../types';
 import { DB } from '../models/index';
 import { v4 as uuidv4 } from 'uuid';
 
+import { NotFoundError } from '../errorTypes';
+
 export class RestaurantDao {
 
 	/**
@@ -12,13 +14,14 @@ export class RestaurantDao {
 		return data;
 	}
 
-
 	/**
 	* @param id
 	*/
 	async get(id: string): Promise<Restaurant | null> {
 		const data = await DB.Models.Restaurant.get(id);
-		data.workingHours = await DB.Models.WorkingHour.list(id);
+		if (data) {
+			data.workingHours = await DB.Models.WorkingHour.list(id);
+		}
 		return data;
 	}
 
@@ -48,8 +51,17 @@ export class RestaurantDao {
 	* @param id
 	*/
 	async update(id: string, restaurant: Restaurant): Promise<Restaurant> {
+		if (!await DB.Models.Restaurant.get(id)) {
+			throw new NotFoundError('This record does not exist. Add it first.')
+		}
+
+		// remove all sub items (add the updated later)
 		await DB.Models.WorkingHour.deleteAll(id);
+
+		// update the main data
 		const data = await DB.Models.Restaurant.update(id, restaurant);
+
+		// add sub items
 		data.workingHours = await Promise.all(restaurant.workingHours.map(hour => {
 			return DB.Models.WorkingHour.add({
 				id: uuidv4(),
